requirement.txt
torch
torchvision
matplotlib
numpy
keras
theano
onnxmltools